Shiny modules providing ui and server logic for export of registry data.

Functions for registries that wants to implement exporting of registry databases, e.g. for local development purposes. Also includes relevant helper functions.

export_ui(id)

export_server(
  id,
  registry_name,
  repo_name = registry_name,
  proxy_url = NULL,
  token = NULL,
  eligible = TRUE
)

export_app(registry_name = "rapbase")

select_list_pubkey(pubkey)

export_db(registry_name, compress = FALSE, session)

Arguments

id

Character string module ID

registry_name

Character string registry name key

repo_name

Character string defining the github repository name of the registry. Default value is registry_name.

proxy_url

Character string providing an internet proxy server url. When set to NULL (default) no proxy server is assumed.

token

Character string representing a GitHub Personal Access Token (PAT) that will be needed if calls to the github api requires authorization. In a development context and if the GITHUB_PAT environmental variable is set token can be left to its default (NULL). In production-like environments token should always be explicitly set.

eligible

Logical defining if the module should be allowed to work at full capacity. This might be useful when access to module products should be restricted. Default is TRUE, i.e. no restrictions.

pubkey

Character vector with public keys

compress

Logical if export data is to be compressed (using gzip). FALSE by default.

session

Shiny session object

Value

Shiny objects, mostly. Helper functions may return other stuff too.

Examples

## client user interface function
ui <- shiny::fluidPage(
  shiny::sidebarLayout(
    shiny::sidebarPanel(
      export_ui("test"),
    ),
    shiny::mainPanel(
      NULL
    )
  )
)

## server function
server <- function(input, output, session) {
  export_server("test", registry_name = "rapbase")
}

## run the shiny app in an interactive environment
if (interactive()) {
  shiny::shinyApp(ui, server)
}